requirements.txt
flask
gunicorn
sqlalchemy
flask_restful
flask_sqlalchemy
psycopg2-binary
requests

